Code refactored following the new UML scheme
In the new UML scheme the GenericActuator class has only the
revpi lib as a field and the state.

All the specifics methods are then refined in the child classes, that
are SingleMotionActuator, DoubleMotionActuator and vacuumActuator.

VacuumActuator is prepared to accept the check of the compressor.
All the child and parent base classes have to be refactored considering
the revpi base.