Modified sets, class methods and dto_dict objects
Now sets and class methods operates changes only if there is an actual
change that make sense to be done, i.e. if the method should modify a
variable to a state that is already present, then the method does not
perform the actual changes. In this way is prevented the multiple
publishing of data on the MQTT topic.