Vacuum Carrier speed configuration set
Now the vacuum carrier has the ability to change his speed in accord to
the received configuration file from MQTT.

The speed can be "Low", "Medium" or "High" and, to simulate the speed
variation, in "Low" and "Medium" configurations the carrier stops in the
middle of the process to decrease its average speed.

In "High" variant, the stop equals to zero.